move load function back to load.py file
