+++ src/ebfloeseg/app.py
@@ -9,9 +9,9 @@ from enum import Enum
 from pathlib import Path
 from typing import Annotated, Optional
 
-import requests
 import typer
 
+from ebfloeseg.load import ImageType, load as load_
 from ebfloeseg.masking import create_land_mask
 from ebfloeseg.preprocess import preprocess, preprocess_b
 
@@ -160,31 +160,6 @@ def process_batch(
             future.result()
 
 
-class ImageType(str, Enum):
-    truecolor = "truecolor"
-    cloud = "cloud"
-    landmask = "landmask"
-
-
-def get_width_height(bbox: str, scale: float):
-    """Get width and height for a bounding box where one pixel corresponds to `scale` bounding box units
-
-    Examples:
-        >>> get_width_height("0,0,1,1", 1)
-        (1, 1)
-
-        >>> get_width_height("0,0,10,50", 5)
-        (2, 10)
-
-    """
-    x1, y1, x2, y2 = [float(n) for n in bbox.split(",")]
-    x_length = abs(x2 - x1)
-    y_length = abs(y2 - y1)
-
-    width, height = int(x_length / scale), int(y_length / scale)
-    return width, height
-
-
 @app.command()
 def load(
     outfile: Annotated[Path, typer.Argument()],
@@ -199,36 +174,19 @@ def load(
     ts: int = 1683675557694,
     format: str = "image/tiff",
 ):
-
-    match kind:
-        case ImageType.truecolor:
-            layers = "MODIS_Terra_CorrectedReflectance_TrueColor"
-        case ImageType.cloud:
-            layers = "MODIS_Terra_Cloud_Fraction_Day"
-        case ImageType.landmask:
-            layers = "OSM_Land_Mask"
-
-    width, height = get_width_height(bbox, scale)
-    _logger.info("Width: %s Height: %s" % (width, height))
-
-    url = f"https://wvs.earthdata.nasa.gov/api/v1/snapshot"
-    payload = {
-        "REQUEST": "GetSnapshot",
-        "TIME": datetime,
-        "BBOX": bbox,
-        "CRS": crs,
-        "LAYERS": layers,
-        "WRAP": wrap,
-        "FORMAT": format,
-        "WIDTH": width,
-        "HEIGHT": height,
-        "ts": ts,
-    }
-    r = requests.get(url, params=payload, allow_redirects=True)
-    r.raise_for_status()
+    result = load_(
+        datetime=datetime,
+        wrap=wrap,
+        kind=kind,
+        bbox=bbox,
+        scale=scale,
+        crs=crs,
+        ts=ts,
+        format=format,
+    )
 
     with open(outfile, "wb") as f:
-        f.write(r.content)
+        f.write(result.content)
 
     return
